Moved debugger scripts to builtins
Always include the debugger scripts to a debug build
