+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/pipeline/GameProjectBuilder.java
@@ -412,6 +412,12 @@ public class GameProjectBuilder extends Builder<Void> {
             }
         }
 
+        // Editor debugger scripts
+        if (project.option("variant", Bob.VARIANT_DEBUG) == Bob.VARIANT_DEBUG) {
+            resources.add(project.getResource("/builtins/scripts/debugger.luac").output().getAbsPath());
+            resources.add(project.getResource("/builtins/scripts/edn.luac").output().getAbsPath());
+            resources.add(project.getResource("/builtins/scripts/mobdebug.luac").output().getAbsPath());
+        }
         return resources;
     }
