+++ engine/engine/content/builtins/scripts/debugger.lua
@@ -1,17 +1,17 @@
 -- Copyright 2020 The Defold Foundation
 -- Licensed under the Defold License version 1.0 (the "License"); you may not use
 -- this file except in compliance with the License.
--- 
+--
 -- You may obtain a copy of the License, together with FAQs at
 -- https://www.defold.com/license
--- 
+--
 -- Unless required by applicable law or agreed to in writing, software distributed
 -- under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 -- CONDITIONS OF ANY KIND, either express or implied. See the License for the
 -- specific language governing permissions and limitations under the License.
 
-local edn      = require '_defold.debugger.edn'
-local mobdebug = require '_defold.debugger.mobdebug'
+local edn      = require 'builtins.scripts.edn'
+local mobdebug = require 'builtins.scripts.mobdebug'
 
 local M = {}
 
@@ -40,7 +40,7 @@ function M.start(port)
   mobdebug.listen(port)
 
   -- for future "break on error" functionality
-  --debug.traceback = function(thread, message, level) 
+  --debug.traceback = function(thread, message, level)
   --	mobdebug.pause()
   --end
 end
